+++ index.js
@@ -7,25 +7,28 @@ var sass = require('node-sass')
 module.exports = SassCompiler
 SassCompiler.prototype = Object.create(CachingWriter.prototype)
 SassCompiler.prototype.constructor = SassCompiler
-function SassCompiler (inputTrees, inputFile, outputFile, options) {
-  if (!(this instanceof SassCompiler)) return new SassCompiler(inputTrees, inputFile, outputFile, options)
-  if (!Array.isArray(inputTrees)) throw new Error('Expected array for first argument - did you mean [tree] instead of tree?')
+function SassCompiler (inputNodes, inputFile, outputFile, options) {
+  if (options == null) options = {}
+  if (!(this instanceof SassCompiler)) return new SassCompiler(inputNodes, inputFile, outputFile, options)
+  if (!Array.isArray(inputNodes)) throw new Error('Expected array for first argument - did you mean [node] instead of node?')
 
-  CachingWriter.call(this, inputTrees, options)
+  CachingWriter.call(this, inputNodes, {
+    annotation: options.annotation
+  })
 
   this.inputFile = inputFile
   this.outputFile = outputFile
-  this.options = options || {}
+  this.options = options
 }
 
 
-SassCompiler.prototype.updateCache = function(includePaths, destDir) {
-  var destFile = path.join(destDir, this.outputFile)
+SassCompiler.prototype.build = function() {
+  var destFile = path.join(this.outputPath, this.outputFile)
   mkdirp.sync(path.dirname(destFile))
 
   var sassOptions = {
-    file: path.join(includePaths[0], this.inputFile),
-    includePaths: includePaths,
+    file: path.join(this.inputPaths[0], this.inputFile),
+    includePaths: this.inputPaths,
 
     imagePath: this.options.imagePath,
     outputStyle: this.options.outputStyle,